Enhance AOT-Support (Issue#3171)
resolve merge conflicts
